src/api/requirements.txt
fastapi[standard]==0.115.6
aiohttp==3.11.9
pydantic==2.9.2
pydantic-settings==2.6.1
nest_asyncio==1.6.0
fastapi-profiler==1.4.1
semantic-kernel[azure]==1.20
requests==2.32.3
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prance==23.6.21.0
python-dotenv==1.0.1
azure-identity==1.19.0
azure-mgmt-resource==23.2.0
azure-monitor-opentelemetry==1.6.4
azure-monitor-opentelemetry-exporter==1.0.0b33